pyproject.toml
[tool.poetry]
name = "webcheck"
version = "0.1.0"
description = "web monitoring"
authors = ["Paul Sorenson <pms@metrak.com>"]
readme = "README.md"
repository = "https://github.com/aiven-recruitment/seniorsre-20210618-paulsorenson"
packages = [
    { include = "__main__.py"},
    { include = "webcheck" }
]
[tool.poetry.dependencies]
python = "^3.9"
aiokafka = "^0.7.1"
aiohttp = "^3.7.4"
asyncpg = "^0.23.0"
typer = "^0.3.2"
PyYAML = "^5.4.1"
munch = "^2.5.0"
keyring = {version = "^23.0.1", optional = true}

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
mypy = "^0.902"
black = "^21.6b0"
pre-commit = "^2.13.0"
types-PyYAML = "^5.4.3"
semgrep = "^0.56.0"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
